rename tof_val_3d
