+++ source/consumers/CMakeLists.txt
@@ -18,7 +18,7 @@ set(SOURCES
   ${dir}/time_domain.cpp
   ${dir}/tof_1d.cpp
   ${dir}/tof_1d_correlate.cpp
-  ${dir}/tof_val_2d.cpp
+  ${dir}/TOFVal2D.cpp
   ${dir}/tof_val_2d_correlate.cpp
   )
 
@@ -36,7 +36,7 @@ set(HEADERS
   ${dir}/time_domain.h
   ${dir}/tof_1d.h
   ${dir}/tof_1d_correlate.h
-  ${dir}/tof_val_2d.h
+  ${dir}/TOFVal2D.h
   ${dir}/tof_val_2d_correlate.h
   )
