+++ source/consumers/consumers_autoreg.cpp
@@ -11,7 +11,7 @@
 #include <consumers/time_domain.h>
 #include <consumers/time_delta_1d.h>
 
-#include <consumers/tof_val_2d.h>
+#include <consumers/TOFVal2D.h>
 #include <consumers/tof_val_2d_correlate.h>
 #include <consumers/histogram_2d.h>
 #include <consumers/image_2d.h>
