+++ tests/consumers/CMakeLists.txt
@@ -13,7 +13,7 @@ set(SOURCES
   ${dir}/time_domain.cpp
   ${dir}/tof_1d.cpp
   ${dir}/tof_1d_correlate.cpp
-  ${dir}/tof_val_2d.cpp
+  ${dir}/TOFVal2DTest.cpp
   ${dir}/tof_val_2d_correlate.cpp
   )
